+++ src/features/MergeNetworks/models/Impl/IntersectionMerge.ts
